structure api based on domains
