+++ src/api/vault/notes/bookmarks/bookmarks.types.ts
@@ -1,0 +1,3 @@
+export type BookmarkFile = { type: "file"; path: string; title: string };
+export type BookMarkGroup = { type: "group"; title: string; items: BookmarkEntry[] };
+export type BookmarkEntry = BookmarkFile | BookMarkGroup;
