+++ src/components/CreateNoteForm.tsx
@@ -1,8 +1,8 @@
 import { ActionPanel, Form, Action, useNavigation, getPreferenceValues, Keyboard } from "@raycast/api";
-
-import NoteCreator from "../utils/data/creator";
-import { FormValue, Vault } from "../utils/interfaces";
-import { renewCache } from "../utils/data/cache";
+import { renewCache } from "../api/cache/cache.service";
+import { createNote } from "../api/vault/notes/notes.service";
+import { CreateNoteParams } from "../api/vault/notes/notes.types";
+import { Vault } from "../api/vault/vault.types";
 import { NoteFormPreferences } from "../utils/preferences";
 
 export function CreateNoteForm(props: { vault: Vault; showTitle: boolean }) {
@@ -40,12 +40,11 @@ export function CreateNoteForm(props: { vault: Vault; showTitle: boolean }) {
     return parsedTags;
   }
 
-  async function createNewNote(noteProps: FormValue, path: string | undefined = undefined) {
+  async function createNewNote(params: CreateNoteParams, path: string | undefined = undefined) {
     if (path !== undefined) {
-      noteProps.path = path;
+      params.path = path;
     }
-    const nc = new NoteCreator(noteProps, vault, pref);
-    const saved = nc.createNote();
+    const saved = createNote(vault, params);
     if (await saved) {
       renewCache(vault);
     }
@@ -61,7 +60,7 @@ export function CreateNoteForm(props: { vault: Vault; showTitle: boolean }) {
           {parseFolderActions()?.map((folder, index) => (
             <Action.SubmitForm
               title={"Create in " + folder}
-              onSubmit={(props: FormValue) => createNewNote(props, folder)}
+              onSubmit={(props: CreateNoteParams) => createNewNote(props, folder)}
               key={index}
               shortcut={{ modifiers: ["shift", "cmd"], key: index.toString() as Keyboard.KeyEquivalent }}
             ></Action.SubmitForm>
