+++ src/components/NoteList/NoteListObsidian.tsx
@@ -1,12 +1,14 @@
 import React, { useReducer } from "react";
 
 import { NotesContext, NotesDispatchContext, useNotes } from "../../utils/hooks";
-import { Note, Vault, SearchArguments } from "../../utils/interfaces";
+import { SearchArguments } from "../../utils/interfaces";
 import { NoteList } from "./NoteList";
 import { NoteActions, OpenNoteActions } from "../../utils/actions";
-import { NoteReducer } from "../../utils/data/reducers";
+import { NoteReducer } from "../../utils/reducers";
 import { Action, Icon } from "@raycast/api";
-import { renewCache } from "../../utils/data/cache";
+import { Vault } from "../../api/vault/vault.types";
+import { Note } from "../../api/vault/notes/notes.types";
+import { renewCache } from "../../api/cache/cache.service";
 
 export function NoteListObsidian(props: {
   vault: Vault;
