+++ src/components/VaultSelection.tsx
@@ -1,9 +1,9 @@
 import { List, ActionPanel, Action } from "@raycast/api";
+import { Vault } from "../api/vault/vault.types";
+import { ReactNode } from "@raycast/api/node_modules/@types/react";
 import { ShowVaultInFinderAction } from "../utils/actions";
 
-import { Vault } from "../utils/interfaces";
-
-export function VaultSelection(props: { vaults: Vault[]; target: (vault: Vault) => any }) {
+export function VaultSelection(props: { vaults: Vault[]; target: (vault: Vault) => ReactNode }) {
   const { vaults, target } = props;
   return (
     <List>
