+++ src/openVaultCommand.tsx
@@ -1,8 +1,9 @@
 import { Action, ActionPanel, closeMainWindow, List, open, popToRoot, Icon } from "@raycast/api";
 
-import { getObsidianTarget, ObsidianTargetType, useObsidianVaults } from "./utils/utils";
+import { getObsidianTarget, ObsidianTargetType } from "./utils/utils";
 import { NoVaultFoundMessage } from "./components/Notifications/NoVaultFoundMessage";
 import { ShowVaultInFinderAction } from "./utils/actions";
+import { useObsidianVaults } from "./utils/hooks";
 
 export default function Command() {
   const { ready, vaults } = useObsidianVaults();
